requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.4
